Move HeapProportionalCacheSizer to org.gradle.cache.internal
