+++ subprojects/core/src/main/java/org/gradle/cache/internal/CacheAccessWorker.java
@@ -16,7 +16,6 @@
 
 package org.gradle.cache.internal;
 
-import org.gradle.api.internal.cache.HeapProportionalCacheSizer;
 import org.gradle.cache.AsyncCacheAccess;
 import org.gradle.cache.CacheAccess;
 import org.gradle.internal.Factory;
