+++ subprojects/launcher/src/main/java/org/gradle/launcher/daemon/configuration/DaemonJvmOptions.java
@@ -17,8 +17,8 @@
 package org.gradle.launcher.daemon.configuration;
 
 import com.google.common.collect.ImmutableSet;
-import org.gradle.api.internal.cache.HeapProportionalCacheSizer;
 import org.gradle.api.specs.Spec;
+import org.gradle.cache.internal.HeapProportionalCacheSizer;
 import org.gradle.internal.file.PathToFileResolver;
 import org.gradle.process.internal.CurrentProcess;
 import org.gradle.process.internal.JvmOptions;
